feat: rewrite entry one by one, reduce the time users wait for merging
